Refactor Turing router to remove overlapping metrics methods
