+++ engines/router/missionctl/fiberapi/interceptors.go
@@ -5,13 +5,15 @@ import (
 	"strings"
 	"time"
 
+	"github.com/caraml-dev/turing/engines/router/missionctl/instrumentation"
 	"github.com/gojek/fiber"
 	"github.com/opentracing/opentracing-go"
 
-	"github.com/caraml-dev/turing/engines/router/missionctl/instrumentation/metrics"
 	"github.com/caraml-dev/turing/engines/router/missionctl/instrumentation/tracing"
 	"github.com/caraml-dev/turing/engines/router/missionctl/log"
 	"github.com/caraml-dev/turing/engines/router/missionctl/turingctx"
+
+	"github.com/gojek/mlp/api/pkg/instrumentation/metrics"
 )
 
 type ctxKey string
@@ -157,7 +159,7 @@ func (i *MetricsInterceptor) AfterCompletion(
 							"traffic_rule": trafficRule,
 						}
 						metrics.Glob().MeasureDurationMsSince(
-							metrics.RouteRequestDurationMs,
+							instrumentation.RouteRequestDurationMs,
 							startTime,
 							labels,
 						)
